backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.3
gunicorn==21.2.0
